+++ CHANGELOG.md
@@ -8,3 +8,12 @@
 - `DateTime` class
 - `isLeapYear` function
 - `leapDays` function
+- `monthsInYear` constant
+- `daysInYear` constant
+- `DaysInMonth` type
+- `daysInMonth` function
+- `daysInWeek` constant
+- `hoursInDay` constant
+- `minutesInHour` constant
+- `secondsInMinute` constant
+- `millisInSecond` constant